Add an ENTRYPOINT instruction to a dockerfile

Adds an ENTRYPOINT instruction to configure a container that will run as an executable.

Usage

dfi_entrypoint(dockerfile, entrypoint)

Arguments

dockerfile

A dockerfile object

entrypoint

Entrypoint command (character vector or string)

Value

An updated dockerfile object with the ENTRYPOINT instruction added

Details

The ENTRYPOINT instruction defines the executable that will be run when the container starts. Any command line arguments passed to docker run will be appended to the entrypoint command.

When used together with CMD, the CMD instruction provides default arguments to the ENTRYPOINT command that can be overridden at runtime.

The function automatically converts the command to JSON array format if a vector is provided, which is the recommended approach for proper signal handling.

See also

Examples

# Simple entrypoint
df <- dockerfile() |>
  dfi_from("rocker/r-ver:4.4.0") |>
  dfi_entrypoint("R")
df
#> FROM rocker/r-ver:4.4.0
#> ENTRYPOINT R 
   
# Array format (recommended)
df <- dockerfile() |>
  dfi_from("rocker/r-ver:4.4.0") |>
  dfi_entrypoint(c("R", "--no-save"))
df
#> FROM rocker/r-ver:4.4.0
#> ENTRYPOINT ["R","--no-save"]