Add a RUN instruction to a dockerfile

Adds a RUN instruction to execute commands in a new layer on top of the current image and commit the results.

Usage

dfi_run(dockerfile, commands)

Arguments

dockerfile

A dockerfile object

commands

Commands to run (character vector for multiple commands)

Value

An updated dockerfile object with the RUN instruction added

Details

When providing multiple commands as a character vector, they will be properly formatted with line continuations in the Dockerfile. This is the recommended approach for package installations to reduce the number of layers in the final image.

See also

Examples

# Single command
df <- dockerfile() |>
  dfi_from("rocker/r-ver:4.4.0") |>
  dfi_run("apt update")
df
#> FROM rocker/r-ver:4.4.0
#> RUN apt update 

# Multiple commands in a single RUN (better practice)
df <- dockerfile() |>
  dfi_from("rocker/r-ver:4.4.0") |>
  dfi_run(c(
    "apt update",
    "apt install -y --no-install-recommends libcurl4-openssl-dev",
    "apt clean",
    "rm -rf /var/lib/apt/lists/*"
  ))