Add a WORKDIR instruction to a dockerfile

Adds a WORKDIR instruction to set the working directory for any subsequent RUN, CMD, ENTRYPOINT, COPY, and ADD instructions.

Usage

dfi_workdir(dockerfile, path)

Arguments

dockerfile

A dockerfile object

path

Working directory path in the container

Value

An updated dockerfile object with the WORKDIR instruction added

Details

The WORKDIR instruction can be used multiple times in a Dockerfile. If a relative path is provided, it will be relative to the previous WORKDIR instruction. If the directory doesn't exist, it will be created.

See also

Examples

df <- dockerfile() |>
  dfi_from("rocker/r-ver:4.4.0") |>
  dfi_workdir("/app") |>
  dfi_copy(".", "/app/")