Create a dockerignore template for {renv}-related files

Creates a dockerignore template with patterns to ignore {renv} library and cache directories.

Usage

dk_template_ignore_renv(.dockerignore = NULL)

Arguments

.dockerignore

Optional existing dockerignore object to add patterns to

Value

A dockerignore object with {renv}-related ignore patterns

Details

This template adds patterns to ignore {renv} library and cache directories, which are typically not needed in a Docker image. These include:

renv/library/: Package library

renv/staging/: Staging area for packages

renv/python/: Python libraries

renv/local/: Local cache

When using {renv} in a Dockerfile, you typically want to copy just the renv.lock file and use renv::restore() to rebuild the library inside the container, rather than copying the entire library.

See also

Examples

# Create a new dockerignore with renv patterns
di <- dk_template_ignore_renv()
di
#> renv/library/
#> renv/staging/
#> renv/python/
#> renv/local/ 

# Add renv patterns to an existing dockerignore
di <- dockerignore() |>
  di_add("*.log") |>
  dk_template_ignore_renv()
di
#> *.log
#> renv/library/
#> renv/staging/
#> renv/python/
#> renv/local/