Print method for dockerfile objects

Displays the contents of a dockerfile object in a readable format, showing each instruction on a new line as it would appear in an actual Dockerfile.

Usage

# S3 method for class 'dockerfile'
print(x, ...)

Arguments

x

A dockerfile object

...

Additional arguments (not used)

Value

Invisibly returns the dockerfile object

See also

Examples

# Create a new dockerfile and add a couple of instructions
df <- dockerfile() |>
  dfi_from("rocker/r-ver:latest") |>
  dfi_run("apt-get update")

# Print the dockerfile
print(df)
#> FROM rocker/r-ver:latest
#> RUN apt-get update